Test: add unit tests for popup.js functions
