+++ src/tests/commentsCounter.test.js
@@ -2,12 +2,12 @@
  * @jest-environment jsdom
  */
 
-import commentsCounter from "../../modules/commentsCounter.js";
-import { fakeComments } from "../../utl/mockData.js";
+import commentsCounter from "../modules/commentsCounter.js";
+import { fakeComments } from "../utl/mockData.js";
 
 describe('Unit Tests', () => {
   describe('commentsCounter()', () => {
-    test('returns number of comments per one card', () => {
+    test('returns number of comments per card', () => {
       expect(commentsCounter(fakeComments)).toEqual(2);
     })
   });
